Duration#to_s: Avoid multiple to_nano calls
Make the 0 entry 0.0 to simplify the type of DIVISORS and avoid calling
both Integer and Float methods in a loop where we don't strictly need
to.

Appease steep's concern over the theoretical nilability of div and unit.
It would be nice to be able to encode that our find call is infallible,
but for now just mention it in a comment.

Also mention what the octothorp does in a float format string to qualify
how the zero-trimming works.